Add iCalendar support
For every sessions/{slug}.html file, we generate a corresponding .ics
file. A new icalendar.ics layout is used to generate an iCalendar file
for every session. On the session HTML page, we can link to the .ics
version. It seems that Jekyll (and hopefully GitHub Pages) will
automatically detect the MIME type and send back a 'text/calendar'
Content-Type header.