Display a modal for choosing Google vs Apple calendar export
This is definitely not an elegant solution and it could do with some CSS
tweaks, but it does degrade gracefully if the user has JavaScript turned
off. By default, the "Add to calendar" button will take the user to the
Google Calendar export. If JavaScript is enabled, we take over the click
handler and use it to display a modal for choosing Google Calendar or
Apple iCal export.